requirements.txt
pandas
scikit-learn
numpy
sentence-transformers
google-generativeai
streamlit
pytest
tqdm